change int to float

+++ src/main/include/Constants.h
@@ -58,7 +58,7 @@ namespace ConstDtm {
 namespace ConstFa {
 
     // Set Arm Motor Speed
-    constexpr int ARM_SPEED = 0.6;
+    constexpr float ARM_SPEED = 0.6;
 
     
 }
